apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
databases[postgresql]==0.8.0 